% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMinIndex}
\alias{getMinIndex}
\title{Get index of minimum value in integer vector}
\usage{
getMinIndex(v)
}
\arguments{
\item{v}{An integer vector.}
}
\description{
Get index of minimum value in integer vector
}
