% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InnerFuzzyDateJoin}
\alias{InnerFuzzyDateJoin}
\title{Inner fuzzy date join}
\usage{
InnerFuzzyDateJoin(x, y, x_id_col, y_id_col, x_date_col, y_date_col,
  x_intvl_less = 0L, x_intvl_more = 0L)
}
\arguments{
\item{x}{X data frame.}

\item{y}{Y data frame.}

\item{x_id_col}{Name of column in X data frame with IDs.}

\item{y_id_col}{Name of column in Y data frame with IDs.}

\item{x_date_col}{Name of column in X data frame with dates.}

\item{y_date_col}{Name of column in Y data frame with dates.}

\item{x_intvl_less}{Number of days before X date to match to Y date. Defaults to 0.}

\item{x_intvl_more}{Number of days after X date to match to Y date. Defaults to 0.}
}
\description{
Inner fuzzy date join
}
