% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_fuzzy_date_join.R
\name{right_fuzzy_date_join}
\alias{right_fuzzy_date_join}
\title{A Function to Perform Right Outer Joins of Longitudinal Data}
\usage{
right_fuzzy_date_join(x, y, x_id_col, y_id_col, x_date_col, y_date_col,
  y_intvl_less = 0, y_intvl_more = 0, keep_x_id = TRUE)
}
\arguments{
\item{x}{X data frame.}

\item{y}{Y data frame.}

\item{x_id_col}{Name of column in X data frame with IDs.}

\item{y_id_col}{Name of column in Y data frame with IDs.}

\item{x_date_col}{Name of column in X data frame with dates.}

\item{y_date_col}{Name of column in Y data frame with dates.}

\item{y_intvl_less}{Number of days before Y date to fuzzy match to X date. Defaults to 0.}

\item{y_intvl_more}{Number of days after Y date to fuzzy match to X date. Defaults to 0.}

\item{keep_x_id}{Keep column in X data frame with IDs? Defaults to TRUE.}
}
\description{
This function allows you to join longitudinal data sets.
}
\examples{
# Define basic data frames X and Y
X <- data.frame(
  x_id = c(1L, 1L, 1L, 2L, 2L, 3L, 3L, 4L, 5L, 5L),
  x_date = as.Date(c("2015-06-01", "2016-06-28", "2017-05-25", "2014-02-23",
                     "2015-03-09", "2016-08-02", "2016-12-13", "2018-03-28",
                     "2017-05-17", "2018-04-03")),
  x_data = runif(10, min = 0, max = 10))

Y <- data.frame(
  y_id = c(1L, 1L, 2L, 2L, 2L, 3L, 4L, 4L, 5L, 5L),
  y_date = as.Date(c("2015-06-01", "2017-05-20", "2014-02-23", "2015-03-10",
                     "2015-03-11", "2016-08-02", "2018-04-01", "2017-03-22",
                     "2017-05-16", "2017-05-17")),
  y_data = runif(10, min = -100, max = 0))

# Use default values where possible
right_fuzzy_date_join(x = X, y = Y,
                      x_id = "x_id", y_id = "y_id",
                      x_date = "x_date", y_date = "y_date")

# Define fuzzy date matching intervals and remove `y_id` column
intvl_less <- 5
intvl_more <- 3
right_fuzzy_date_join(x = X, y = Y,
                      x_id = "x_id", y_id = "y_id",
                      x_date = "x_date", y_date = "y_date",
                      y_intvl_less = intvl_less, y_intvl_more = intvl_more,
                      keep_x_id = FALSE)
}
\keyword{data,}
\keyword{fuzzy,}
\keyword{join,}
\keyword{longitudinal}
\keyword{right,}
